made anime card component reusable
